⬆ libv5rts@1.0.0b20
Summary:
🎉 100 diffs 🎉

- Upgrade libv5rts to beta 20
- Adjust new parameter types per SDK differences
  - Some migrations to `int32_t`s
  - Using millivolts and milliamps in places where it makes sense
  - We aren't changing some of our public facing API to reflect the internal changes (most notable ADI port changes)

Test Plan:
[X] Clean compile
[X] Test basic_test.c
[X] Tested some code at Worlds
[X] Verify symbols

Reviewers: O1 The Ori, brookea

Reviewed By: O1 The Ori, brookea

Subscribers: baylessj, brookea

Tags: #zorp

Differential Revision: http://phabricator.purduesigbots.com/D100

+++ include/pros/misc.h
@@ -20,6 +20,8 @@
 #ifndef _PROS_MISC_H_
 #define _PROS_MISC_H_
 
+#include <stdint.h>
+
 #define NUM_V5_PORTS (22)
 
 /******************************************************************************/
@@ -185,7 +187,7 @@ int32_t controller_get_digital_new_press(controller_id_e_t id, controller_digita
  *
  * \return The current voltage of the battery
  */
-double battery_get_voltage(void);
+int32_t battery_get_voltage(void);
 
 /**
  * Gets the current current of the battery, as reported by VEXos.
@@ -195,7 +197,7 @@ double battery_get_voltage(void);
  *
  * \return The current current of the battery
  */
-double battery_get_current(void);
+int32_t battery_get_current(void);
 
 /**
  * Gets the current temperature of the battery, as reported by VEXos.
